+++ lib/smart_core/types.rb
@@ -4,6 +4,7 @@ require 'smart_core'
 
 # @api public
 # @since 0.1.0
+# @version 0.4.0
 module SmartCore::Types
   require_relative 'types/version'
   require_relative 'types/errors'
@@ -11,5 +12,6 @@ module SmartCore::Types
   require_relative 'types/primitive'
   require_relative 'types/value'
   require_relative 'types/protocol'
+  require_relative 'types/struct'
   require_relative 'types/variadic'
 end