+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.4.0] - 2020-12-22
+### Added
+- Brand new `SmartCore::Types::Variadic::Enum` type (runtime-based type);
+- Brand new `SmartCore::Types::Struct` type category and new types:
+  - `SmartCore::Types::Struct::StrictHash` (runtime-based type);
+
 ## [0.3.0] - 2020-12-22
 ### Added
 - Extended **Type Definition API**: support for **runtime attributes**: